convert ItemsLegend.spec.tsx to RTL
